refactor: common 패키지명을 support로 수정
- extensions 패키지 이동
