+++ src/main/kotlin/nexters/admin/support/auth/AuthAnnotations.kt
@@ -1,4 +1,4 @@
-package nexters.admin.common.auth
+package nexters.admin.support.auth
 
 @Target(AnnotationTarget.VALUE_PARAMETER)
 @Retention(AnnotationRetention.RUNTIME)
