+++ src/main/kotlin/nexters/admin/support/exception/ExceptionControllerAdvice.kt
@@ -1,4 +1,4 @@
-package nexters.admin.common.exception
+package nexters.admin.support.exception
 
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
